Update README.md
Hello there, I'm Femi! 👋 I'm a passionate frontend web developer with a strong understanding of web development and front-end technologies. I'm well-versed in responsive design principles and have experience optimizing websites for performance and accessibility. With a keen eye for detail, I always strive to create visually appealing and intuitive user interfaces.

In addition, I'm a fast learner with excellent communication skills, which makes me highly effective when working with teams. I have an in-depth and practical understanding of Object-Oriented Programming, MVC architecture, and the Publisher Subscriber Pattern along with their principles. So far, I've had the opportunity to work on several amazing projects, some of which are live and can be viewed in the projects section of my portfolio.

I'm currently open to full-time roles and would be eager to participate in a scheduled interview at any time.